MyCalculator
MyCalculator is a C# desktop/mobile application that serves several
functions:  The first is to be used as a fully customizable scientific
calculator.  The second is to serve as a tutor to both teach
children/young adults how to use a scientific calculator, as well as to
teach them the high-level math that is behind the calculators’
functionality.  These tutorials will include the explanations of the
high-level math, how it’s done, and a quiz at the end to test the
users.  The quizzes are completely randomized, and the user will
literally never get the same question twice.  Also, the app will store
the users’ high score from every subject